requirements.txt
alembic==1.8.0
apache-airflow==1.10.9
numpy==1.22.4
pandas==0.25.3
scikit-learn==1.1.1